+++ tt_tools_common/utils_common/system_utils.py
@@ -132,13 +132,13 @@ def system_compatibility() -> dict:
     else:
         checklist["Memory"] = (False, "Recommended 32GB+")
 
-    # Due do how ARM pcie device rescans are handled, we can't perform a wormhole reset on ARM systems
+    # Due do how Arm pcie device rescans are handled, we can't perform a wormhole reset on Arm systems
     if host_info["Platform"].startswith("arm") or host_info["Platform"].startswith(
         "aarch"
     ):
         checklist["WH Reset"] = (
             False,
-            "Not supported on ARM",
+            "Not supported on Arm",
         )
     else:
         checklist["WH Reset"] = (True, "Pass")